Large refactoring in progress; full details in pre-merge commit
