--- pyeo/apps/aws_processing/change_detection_aws.ini
@@ -1,36 +1,0 @@
-# Edit the below variables and save
-
-[sent_2]
-#Your username and password for the Coperniucs hub. It's free to register.
-user=
-pass=
-url=https://scihub.copernicus.eu/dhus
-
-[planet]
-#Your API key for Planet, if you have one.
-api_key=
-
-[forest_sentinel]
-
-# aoi_name: The name of this area of interest. No spaces.
-aoi_name=aws_instance
-# Dates in the form yyyymmdd
-start_date=20180304
-end_date=20180404
-# Dates to download and preprocess for the initial cloud-free composite
-compsite_start=20180101
-composite_end=20180301
-# Cloud cover imagery to download
-cloud_cover=30
-# Certainty value above which a pixel is considered a cloud from sen2cor
-cloud_certainty_threshold=50
-
-# All below are constant for AWS implementation
-root_dir=/home/ubuntu/data
-aoi_path=/home/ubuntu/config/aoi.json
-log_path=/home/ubuntu/composite.log
-model=/home/ubuntu/config/model.pkl
-
-[sen2cor]
-# Path to the sen2cor preprocessor script, L2A_Process. Usually in the bin/ folder of your sen2cor installation.
-path=/home/ubuntu/tools/Sen2Cor-02.05.05-Linux64/bin/L2A_Process
