+++ pyeo/apps/change_detection/image_comparison.py
@@ -1,8 +1,12 @@
 """Quick script that stacks and classifies two images"""
 
 import os, sys
+
+import pyeo.classification
+import pyeo.raster_manipulation
+import pyeo.filesystem_utilities
+
 sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..', '..')))
-import pyeo.core as pyeo
 import argparse
 import os
 from tempfile import TemporaryDirectory
@@ -20,14 +24,14 @@ if __name__ == "__main__":
     parser.add_argument("-m", "--mask", action="store_true")
     args = parser.parse_args()
 
-    log = pyeo.init_log(args.log_path)
+    log = pyeo.filesystem_utilities.init_log(args.log_path)
 
     with TemporaryDirectory() as td:
         stacked_path = os.path.join(td, "stacked.tif")
-        pyeo.stack_images([args.old_image, args.new_image], stacked_path, geometry_mode="intersect")
+        pyeo.raster_manipulation.stack_images([args.old_image, args.new_image], stacked_path, geometry_mode="intersect")
 
-        pyeo.classify_image(stacked_path, args.model, args.output, prob_out_path=None,
-                            num_chunks=args.chunks, apply_mask=args.mask)
+        pyeo.classification.classify_image(stacked_path, args.model, args.output, prob_out_path=None,
+                                           num_chunks=args.chunks, apply_mask=args.mask)
