+++ pyeo/apps/change_detection/rolling_composite_s2_change_detection.py
@@ -13,9 +13,13 @@ For each preprocessed image:
 Step 8:
  """
 
-import os, sys
-sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..', '..')))
-import pyeo.core as pyeo
+import pyeo.classification
+import pyeo.preprocessing
+import pyeo.queries_and_downloads
+import pyeo.raster_manipulation
+import pyeo.sen2_funcs
+import pyeo.filesystem_utilities
+
 import configparser
 import argparse
 import os
@@ -88,8 +92,8 @@ if __name__ == "__main__":
     composite_end_date = conf['forest_sentinel']['composite_end']
     epsg = int(conf['forest_sentinel']['epsg'])
 
-    pyeo.create_file_structure(project_root)
-    log = pyeo.init_log(log_path)
+    pyeo.filesystem_utilities.create_file_structure(project_root)
+    log = pyeo.filesystem_utilities.init_log(log_path)
 
     l1_image_dir = os.path.join(project_root, r"images/L1")
     l2_image_dir = os.path.join(project_root, r"images/L2")
@@ -109,7 +113,7 @@ if __name__ == "__main__":
 
     if args.start_date == "LATEST":
         # This isn't nice, but returns the yyyymmdd string of the latest stacked image
-        start_date = pyeo.get_image_acquisition_time(pyeo.sort_by_timestamp(
+        start_date = pyeo.sen2_funcs.get_image_acquisition_time(pyeo.filesystem_utilities.sort_by_timestamp(
             [image_name for image_name in os.listdir(stacked_image_dir) if image_name.endswith(".tif")],
             recent_first=True
         )[0]).strftime("%Y%m%d")
@@ -125,49 +129,49 @@ if __name__ == "__main__":
         if args.do_download or do_all:
             log.info("Downloading for initial composite between {} and {} with cloud cover <= ()".format(
                 composite_start_date, composite_end_date, cloud_cover))
-            composite_products = pyeo.check_for_s2_data_by_date(aoi_path, composite_start_date, composite_end_date,
-                                                             conf, cloud_cover=cloud_cover)
+            composite_products = pyeo.queries_and_downloads.check_for_s2_data_by_date(aoi_path, composite_start_date, composite_end_date,
+                                                                                      conf, cloud_cover=cloud_cover)
             if args.download_l2_data:
                 log.info("Filtering query results for matching L1 and L2 products")
-                composite_products = pyeo.filter_non_matching_s2_data(composite_products)
+                composite_products = pyeo.queries_and_downloads.filter_non_matching_s2_data(composite_products)
                 log.info("{} products remain".format(len(composite_products)))
-            pyeo.download_s2_data(composite_products, composite_l1_image_dir, composite_l2_image_dir,
-                                  source=args.download_source, user=sen_user, passwd=sen_pass, try_scihub_on_fail=True)
+            pyeo.queries_and_downloads.download_s2_data(composite_products, composite_l1_image_dir, composite_l2_image_dir,
+                                                        source=args.download_source, user=sen_user, passwd=sen_pass, try_scihub_on_fail=True)
         if args.do_preprocess or do_all and not args.download_l2_data:
             log.info("Preprocessing composite products")
-            pyeo.atmospheric_correction(composite_l1_image_dir, composite_l2_image_dir, sen2cor_path,
-                                        delete_unprocessed_image=False)
+            pyeo.preprocessing.atmospheric_correction(composite_l1_image_dir, composite_l2_image_dir, sen2cor_path,
+                                                      delete_unprocessed_image=False)
         if args.do_merge or do_all:
             log.info("Aggregating composite layers")
-            pyeo.preprocess_sen2_images(composite_l2_image_dir, composite_merged_dir, composite_l1_image_dir,
-                                        cloud_certainty_threshold, epsg=epsg, buffer_size=5)
+            pyeo.sen2_funcs.preprocess_sen2_images(composite_l2_image_dir, composite_merged_dir, composite_l1_image_dir,
+                                                   cloud_certainty_threshold, epsg=epsg, buffer_size=5)
         log.info("Building initial cloud-free composite")
-        pyeo.composite_directory(composite_merged_dir, composite_dir, generate_date_images=True)
+        pyeo.raster_manipulation.composite_directory(composite_merged_dir, composite_dir, generate_date_images=True)
 
     # Query and download all images since last composite
     if args.do_download or do_all:
-        products = pyeo.check_for_s2_data_by_date(aoi_path, start_date, end_date, conf, cloud_cover=cloud_cover)
+        products = pyeo.queries_and_downloads.check_for_s2_data_by_date(aoi_path, start_date, end_date, conf, cloud_cover=cloud_cover)
         if args.download_l2_data:
             log.info("Filtering query results for matching L1 and L2 products")
-            products = pyeo.filter_non_matching_s2_data(products)
+            products = pyeo.queries_and_downloads.filter_non_matching_s2_data(products)
             log.info("{} products remain".format(len(products)))
         log.info("Downloading")
-        pyeo.download_s2_data(products, l1_image_dir, l2_image_dir, args.download_source, user=sen_user, passwd=sen_pass, try_scihub_on_fail=True)
+        pyeo.queries_and_downloads.download_s2_data(products, l1_image_dir, l2_image_dir, args.download_source, user=sen_user, passwd=sen_pass, try_scihub_on_fail=True)
 
     # Atmospheric correction
     if args.do_preprocess or do_all and not args.download_l2_data:
         log.info("Applying sen2cor")
-        pyeo.atmospheric_correction(l1_image_dir, l2_image_dir, sen2cor_path, delete_unprocessed_image=False)
+        pyeo.preprocessing.atmospheric_correction(l1_image_dir, l2_image_dir, sen2cor_path, delete_unprocessed_image=False)
 
     # Aggregating layers into single image
     if args.do_merge or do_all:
         log.info("Aggregating layers")
-        pyeo.preprocess_sen2_images(l2_image_dir, merged_image_dir, l1_image_dir, cloud_certainty_threshold, epsg=epsg,
-                                    buffer_size=5)
+        pyeo.sen2_funcs.preprocess_sen2_images(l2_image_dir, merged_image_dir, l1_image_dir, cloud_certainty_threshold, epsg=epsg,
+                                               buffer_size=5)
 
     log.info("Finding most recent composite")
     latest_composite_name = \
-        pyeo.sort_by_timestamp(
+        pyeo.filesystem_utilities.sort_by_timestamp(
             [image_name for image_name in os.listdir(composite_dir) if image_name.endswith(".tif")],
             recent_first=True
         )[0]
@@ -176,7 +180,7 @@ if __name__ == "__main__":
 
     log.info("Sorting image list")
     images = \
-        pyeo.sort_by_timestamp(
+        pyeo.filesystem_utilities.sort_by_timestamp(
             [image_name for image_name in os.listdir(merged_image_dir) if image_name.endswith(".tif")],
             recent_first=False
         )
@@ -190,10 +194,10 @@ if __name__ == "__main__":
 
         # Stack with preceding composite
         if args.do_stack or do_all:
-            latest_composite_path = pyeo.get_preceding_image_path(new_image_path, composite_dir)
+            latest_composite_path = pyeo.filesystem_utilities.get_preceding_image_path(new_image_path, composite_dir)
             log.info("Stacking {} with composite {}".format(new_image_path, latest_composite_path))
-            new_stack_path = pyeo.stack_image_with_composite(new_image_path, latest_composite_path, stacked_image_dir,
-                                                             invert_stack=args.flip_stacks)
+            new_stack_path = pyeo.raster_manipulation.stack_image_with_composite(new_image_path, latest_composite_path, stacked_image_dir,
+                                                                                 invert_stack=args.flip_stacks)
         #else new_stack_path =
 
         # Classify with composite
@@ -204,15 +208,16 @@ if __name__ == "__main__":
                 new_prob_image = os.path.join(probability_image_dir, "prob_{}".format(os.path.basename(new_stack_path)))
             else:
                 new_prob_image = None
-            pyeo.classify_image(new_stack_path, model_path, new_class_image, new_prob_image, num_chunks=10,
-                                skip_existing=True, apply_mask=True)
+            pyeo.classification.classify_image(new_stack_path, model_path, new_class_image, new_prob_image, num_chunks=10,
+                                               skip_existing=True, apply_mask=True)
 
         # Build new composite
         if args.do_update or do_all:
             log.info("Updating composite")
             new_composite_path = os.path.join(
-                composite_dir, "composite_{}.tif".format(pyeo.get_sen_2_image_timestamp(os.path.basename(image))))
-            pyeo.composite_images_with_mask(
+                composite_dir, "composite_{}.tif".format(
+                    pyeo.sen2_funcs.get_sen_2_image_timestamp(os.path.basename(image))))
+            pyeo.raster_manipulation.composite_images_with_mask(
                 (latest_composite_path, new_image_path), new_composite_path, generate_date_image=True)
             latest_composite_path = new_composite_path
