+++ pyeo/apps/change_detection/simple_s2_change_detection.py
@@ -13,8 +13,14 @@ $ python pyeo/apps/change_detection/simple_s2_change_detection.py --conf /path/t
 Produces two directories of un-mosaiced imagery; one of classified images and one of class probabilites"""
 
 import os, sys
-sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..', '..')))
-import pyeo.core as pyeo
+
+import pyeo.classification
+import pyeo.preprocessing
+import pyeo.queries_and_downloads
+import pyeo.raster_manipulation
+import pyeo.sen2_funcs
+import pyeo.filesystem_utilities
+
 import configparser
 import argparse
 import os
@@ -54,8 +60,8 @@ if __name__ == "__main__":
     model_path = conf['forest_sentinel']['model']
     sen2cor_path = conf['sen2cor']['path']
 
-    pyeo.create_file_structure(project_root)
-    log = pyeo.init_log(log_path)
+    pyeo.filesystem_utilities.create_file_structure(project_root)
+    log = pyeo.filesystem_utilities.init_log(log_path)
 
     l1_image_path = os.path.join(project_root, r"images/L1")
     l2_image_path = os.path.join(project_root, r"images/L2")
@@ -68,37 +74,37 @@ if __name__ == "__main__":
 
     # Query and download
     if args.do_download or do_all:
-        products = pyeo.check_for_s2_data_by_date(aoi_path, start_date, end_date, conf)
+        products = pyeo.queries_and_downloads.check_for_s2_data_by_date(aoi_path, start_date, end_date, conf)
         log.info("Downloading")
-        pyeo.download_s2_data(products, l1_image_path)
+        pyeo.queries_and_downloads.download_s2_data(products, l1_image_path)
 
     # Atmospheric correction
     if args.do_preprocess or do_all:
         log.info("Applying sen2cor")
-        pyeo.atmospheric_correction(l1_image_path, l2_image_path, sen2cor_path, delete_unprocessed_image=False)
+        pyeo.preprocessing.atmospheric_correction(l1_image_path, l2_image_path, sen2cor_path, delete_unprocessed_image=False)
 
     # Merging / Aggregating layers into single image
     if args.do_merge or do_all:
         log.info("Cleaning L2A directory")
-        pyeo.clean_l2_dir(l2_image_path, resolution="10m", warning=False)
+        pyeo.filesystem_utilities.clean_l2_dir(l2_image_path, resolution="10m", warning=False)
         log.info("Aggregating layers")
-        pyeo.preprocess_sen2_images(l2_image_path, merged_image_path, cloud_certainty_threshold)
+        pyeo.sen2_funcs.preprocess_sen2_images(l2_image_path, merged_image_path, cloud_certainty_threshold)
 
     # Stack layers
     if args.do_stack or do_all:
         log.info("Stacking before and after images")
-        pyeo.create_new_stacks(merged_image_path, stacked_image_path)
+        pyeo.raster_manipulation.create_new_stacks(merged_image_path, stacked_image_path)
 
     # Mosaic stacked layers
     if args.do_stack or do_all:
         log.info("Mosaicking stacked multitemporal images across tiles")
-        pyeo.mosaic_images(stacked_image_path, mosaic_image_path, format="GTiff", datatype=gdal.GDT_Int32, nodata=0)
+        pyeo.raster_manipulation.mosaic_images(stacked_image_path, mosaic_image_path, format="GTiff", datatype=gdal.GDT_Int32, nodata=0)
 
     # Classify stacks
     if args.do_classify or do_all:
         log.info("Classifying images")
-        pyeo.classify_directory(stacked_image_path, model_path, catagorised_image_path, probability_image_path,
-                                num_chunks=16, apply_mask=False)
+        pyeo.classification.classify_directory(stacked_image_path, model_path, catagorised_image_path, probability_image_path,
+                                               num_chunks=16, apply_mask=False)
