--- pyeo/apps/historical_data_processing/date_ranges.txt
@@ -1,3 +1,0 @@
-20170101 20170227
-20170301 20170531
-20170601 20171231
