--- pyeo/apps/historical_data_processing/hist_test.qsub
@@ -1,13 +1,0 @@
-#PBS -o output.txt
-#PBS -e error.txt
-#PBS -q devel
-#PBS -n ForestSentinelTest
-#PBS -l walltime=02:00:00
-#PBS -l vmem=32gb
-#PBS -m bea
-#PBS -M jfr10@le.ac.uk
-
-source activate senenv
-cd /scratch/forest2020/jfr10/ForestSentinel/src/demos_and_apps/historical_data_processing
-export PBS_ARRAYID=0
-python download_and_preprocess_historical_data.py
