--- pyeo/apps/historical_data_processing/historical.conf.template
@@ -1,11 +1,0 @@
-[sent_2]
-user=
-pass=
-
-[forest_sentinel]
-root_dir=
-aoi=
-start_date=
-end_date=
-cloud_cover=
-log_path=
