+++ pyeo/apps/masking/filter_by_class_map.py
@@ -1,8 +1,10 @@
-"""Quick script that stacks and classifies two images"""
-
+"""Filters an image by a given class."""
 import os, sys
+
+import pyeo.filesystem_utilities
+import pyeo.raster_manipulation
+
 sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..', '..')))
-import pyeo.core as pyeo
 import argparse
 import os
 from tempfile import TemporaryDirectory
@@ -18,6 +20,6 @@ if __name__ == "__main__":
     parser.add_argument("-l", "--log_path", default=os.path.join(os.getcwd(), "comparison.log"))
     args = parser.parse_args()
 
-    log = pyeo.init_log(args.log_path)
+    log = pyeo.filesystem_utilities.init_log(args.log_path)
 
-    pyeo.filter_by_class_map(args.input_image, args.class_image, args.output, args.filter_classes)
+    pyeo.raster_manipulation.filter_by_class_map(args.input_image, args.class_image, args.output, args.filter_classes)
