+++ pyeo/apps/subprocessing/extract_signatures.py
@@ -3,12 +3,9 @@
     python extract_signatures.py in_ras ras1.tif ras2.tif out sigs.csv
  """
 
-import os, sys
-sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..', '..')))
-import pyeo.core as pc
+import pyeo.classification
 import csv
 import os
-import glob
 import numpy as np
 import argparse
 
@@ -23,7 +20,7 @@ if __name__ == "__main__":
         training_image_folder, training_image_name = os.path.split(training_image_file_path)
         training_image_name = training_image_name[:-4]  # Strip the file extension
         shape_path = os.path.join(training_image_folder, training_image_name, training_image_name + '.shp')
-        this_training_data, this_classes = pc.get_training_data(training_image_file_path, shape_path)
+        this_training_data, this_classes = pyeo.classification.get_training_data(training_image_file_path, shape_path)
 
         sigs=np.vstack((this_classes, this_training_data.T))
