+++ pyeo/apps/validation/generate_stratified_sample_points.py
@@ -1,10 +1,8 @@
 """App to generate a set of proportional stratified sample points from a class map"""
 
-import sys, os
-sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..', '..')))
 import argparse
 from pyeo.validation import produce_stratifed_validation_points
-from pyeo.core import init_log
+from pyeo.filesystem_utilities import init_log
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser("Generates a shapefile containing n stratified random points"
