--- pyeo/conf.ini
@@ -1,22 +1,0 @@
-[sen2]
-user =
-pass =
-
-[query]
-#aoi_format: can be shapefile or geojson
-aoi_format = shapefile
-#aoi_path: Path to either the .json or the .shp file
-aoi_path = /path/to/aoi
-#Start and end dates of query in the format yyyymmdd
-start_date = 20171001
-end_date = 20171201
-#Maximum cloud cover %
-max_cloud_cover = 50
-
-[data]
-
-out_folder = .
-
-[post_processing]
-#Path to the folder containing L2A_Process.bat
-sen2cor_path = C:\programs\Sen2Cor-02.05.05-win64
